Testrun stabilization

+++ src/main/java/api/middle/iteration2/models/AccountResponseModel.java
@@ -2,12 +2,13 @@ package api.middle.iteration2.models;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class AccountResponseModel {
     private Long id;
     private String accountNumber;
-    private float balance;
+    private BigDecimal balance;
     private List<Transaction> transactions;
 }
